Move remaining stories
